git-svn-id: svn://182.53.32.18/bbip/project/GDEUPSB/TRUNK/GDEUPSB@49166 2ac714ae-9a16-11e4-ad83-2f06aba1930a

+++ src/com/bocom/bbip/gdeupsb/action/zh/AfterBatchAcpServiceImplZHAG00.java
@@ -58,7 +58,7 @@ public class AfterBatchAcpServiceImplZHAG00 extends BaseAction implements AfterB
         	Map<String , Object > map=new HashMap<String, Object>();
         	String sqn=eupsBatchInfoDetail.getRmk1();
         	map.put("sqn", sqn);
-        	String txnAmt=eupsBatchInfoDetail.getTxnAmt().scaleByPowerOfTen(2)+"";
+        	String txnAmt=eupsBatchInfoDetail.getTxnAmt().scaleByPowerOfTen(2).intValue()+"";
         	if(txnAmt.length()<12){
         			txnAmt="0"+txnAmt;
         	}
